pyproject.toml
[project]
name = "hls-nextgen-orchestration"
version = "0.1.0"
description = "HLS nextgen orchestration CDK stack"
authors = [
  {name = "Chris Holden", email = "chris@developmentseed.org"},
  {name = "Sean Harkins", email = "sean@developmentseed.org"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-aws-lambda-python-alpha>=2.191.0a0",
    "aws-cdk-lib>=2.188.0",
    "constructs>=10.4.2",
]

[tool.uv.workspace]
members = ["src/common"]

[tool.uv.sources]
common = { workspace = true }

[dependency-groups]
dev = [
  "aws-lambda-typing>=2.20.0",
  "boto3-stubs[batch,s3,sqs]>=0.37.36",
  "common",
  "moto>=5.1.3",
  "mypy>=1.15.0",
  "nodeenv",
  "pre-commit",
  "pytest>=8.3.5",
  "pytest-cov>=6.1.1",
  "pytest-lazy-fixtures>=1.1.2",
  "pydantic-settings>=2.8.1",
  "responses>=0.25.7",
  "ruff>=0.11.4",
  "types-requests>=2.32.0.20250515",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common", "src/job_monitor", "src/job_requeuer"]

[tool.ruff]
line-length = 88
target-version = "py38"
exclude = ["cdk.out", ".venv", "node_modules"]

[tool.ruff.lint]
select = ["E", "F", "I"]
per-file-ignores = {}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = [
  "common",
  "job_monitor",
  "job_requeuer",
]
combine-as-imports = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true
files = ["cdk", "src", "tests"]
mypy_path = ["cdk", "src"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "s3pathlib"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "boto_session_manager"
ignore_missing_imports = true
